refactor: Update imports in filters.dart for SVG assets
