+++ lib/assets/svg/switch_svg.dart
